Decompose projection file
