+++ crates/cubek-tile/src/physical/projection/mod.rs
@@ -1,0 +1,30 @@
+//! How an operand's logical [`Space`](crate::Space) axes map onto its buffer's
+//! [`PhysicalAxis`](crate::PhysicalAxis) positions.
+//!
+//! The default is one logical axis per physical axis with coefficient `1`: logical rank equals
+//! physical rank, one tile coordinate moves one physical axis by one sub-tile, and sibling windows
+//! never overlap. Matmul fits because `K` genuinely is a physical axis of `lhs` and `rhs`.
+//!
+//! A gather-reduce over an *abstract* dimension does not fit. A stencil's reduce axes address the
+//! same input physical axis its output axes address, each with its own coefficient:
+//!
+//! ```text
+//! space {Oh, Ow, Cout, Rh, Rw, Cin}
+//!   input physical axis Ih <- Oh*stride_h + Rh*dilation_h
+//!         physical axis Iw <- Ow*stride_w + Rw*dilation_w
+//!         physical axis Cin
+//! ```
+//!
+//! So a physical axis is an affine combination of logical axes, and consecutive windows along `Oh`
+//! overlap by the receptive field. [`Projection`] is that combination;
+//! [`direct`](Projection::direct) is the degenerate one every current operand uses.
+
+mod fold;
+mod map;
+mod projection;
+mod tiling;
+
+pub use fold::*;
+pub use map::*;
+pub use projection::*;
+pub use tiling::*;
