+++ crates/cubek-tile/src/physical/projection/projection.rs
@@ -1,219 +1,9 @@
-//! How an operand's logical [`Space`](crate::Space) axes map onto its buffer's
-//! [`PhysicalAxis`](crate::PhysicalAxis) positions.
-//!
-//! The default is one logical axis per physical axis with coefficient `1`: logical rank equals
-//! physical rank, one tile coordinate moves one physical axis by one sub-tile, and sibling windows
-//! never overlap. Matmul fits because `K` genuinely is a physical axis of `lhs` and `rhs`.
-//!
-//! A gather-reduce over an *abstract* dimension does not fit. A stencil's reduce axes address the
-//! same input physical axis its output axes address, each with its own coefficient:
-//!
-//! ```text
-//! space {Oh, Ow, Cout, Rh, Rw, Cin}
-//!   input physical axis Ih <- Oh*stride_h + Rh*dilation_h
-//!         physical axis Iw <- Ow*stride_w + Rw*dilation_w
-//!         physical axis Cin
-//! ```
-//!
-//! So a physical axis is an affine combination of logical axes, and consecutive windows along `Oh`
-//! overlap by the receptive field. [`Projection`] is that combination;
-//! [`direct`](Projection::direct) is the degenerate one every current operand uses.
+//! An operand's logical axes mapped onto its buffer's physical axes: the affine combination the
+//! module doc derives, assembled from one [`PhysicalAxisMap`] per physical axis.
 
-use cubecl::prelude::*;
-use cubecl::std::tensor::layout::CoordsDyn;
 use cubecl::zspace::SmallVec;
 
-use crate::{Axis, ConcreteLayout, Coords, Fold, FoldExpand, FoldSeq, FoldSeqExpand, MAX_AXES};
-
-/// How far one unit of a logical axis's coordinate moves along one physical axis. Mirrors
-/// [`Extent`](crate::Extent): `Static` is a comptime constant so the advance folds the way
-/// [`window_start`](crate::MemData) needs, `Dynamic` is reserved for a runtime stride/dilation
-/// and is rejected by [`Projection::validate`] until the runtime half exists.
-#[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
-pub enum Scale {
-    Static(usize),
-    Dynamic,
-}
-
-impl Scale {
-    /// The comptime coefficient; panics on `Dynamic`.
-    pub fn get(self) -> usize {
-        match self {
-            Scale::Static(n) => n,
-            Scale::Dynamic => {
-                panic!(
-                    "Scale::get: this coefficient is Dynamic; its value is only known at runtime"
-                )
-            }
-        }
-    }
-
-    pub fn is_dynamic(self) -> bool {
-        matches!(self, Scale::Dynamic)
-    }
-}
-
-/// One logical axis's contribution to one physical axis: `digit * scale`, where the digit is the
-/// whole coordinate unless the axis is spread over several physical axes, which
-/// [`Projection::digit`] reads off the map's own shape rather than off a stored constant.
-#[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
-pub struct AxisTerm {
-    pub axis: Axis,
-    pub scale: Scale,
-}
-
-/// One [`PhysicalAxis`](crate::PhysicalAxis) as an affine combination of logical axes' digits:
-/// `physical = Σ digit(axis) * scale`.
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct PhysicalAxisMap {
-    terms: SmallVec<[AxisTerm; MAX_AXES]>,
-}
-
-impl PhysicalAxisMap {
-    /// The identity map: this physical axis *is* `axis`, coefficient `1`. What every operand of a
-    /// non-gather operation uses on every physical axis.
-    pub fn of(axis: Axis) -> Self {
-        PhysicalAxisMap {
-            terms: SmallVec::from_slice(&[AxisTerm {
-                axis,
-                scale: Scale::Static(1),
-            }]),
-        }
-    }
-
-    /// An affine combination, e.g. `affine(&[(Oh, stride), (Rh, dilation)])`.
-    pub fn affine(terms: &[(Axis, usize)]) -> Self {
-        PhysicalAxisMap {
-            terms: terms
-                .iter()
-                .map(|&(axis, scale)| AxisTerm {
-                    axis,
-                    scale: Scale::Static(scale),
-                })
-                .collect(),
-        }
-    }
-
-    pub fn terms(&self) -> &[AxisTerm] {
-        &self.terms
-    }
-
-    /// `axis`'s coefficient, `0` when it does not address this physical axis.
-    pub fn scale(&self, axis: Axis) -> usize {
-        self.terms
-            .iter()
-            .find(|t| t.axis == axis)
-            .map_or(0, |t| t.scale.get())
-    }
-
-    /// Whether this physical axis is exactly `axis` at coefficient `1`. Says nothing about digit
-    /// extraction, which is a property of the whole [`Projection`] (how many physical axes carry
-    /// `axis`), not of one map.
-    pub fn is_identity(&self, axis: Axis) -> bool {
-        matches!(
-            self.terms.as_slice(),
-            [AxisTerm {
-                axis: a,
-                scale: Scale::Static(1)
-            }] if *a == axis
-        )
-    }
-}
-
-/// How many physical fragments each logical axis is split across, in the operand's own axis order.
-/// One fragment is an untiled axis; `n` fragments make a coordinate along it an `n`-digit mixed
-/// radix number ([`Projection::digit`]), which is the whole encoding of storage tiling. No extent
-/// appears here: the radices are read off the buffer's `physical_shape` at use time.
-///
-/// The physical order this induces is level-major, coarsest first: every axis contributes its
-/// level-0 fragment, then every axis still deep enough contributes its level-1 fragment, down to
-/// the tile fragments. For the common case of untiled leading axes over a uniformly tiled block
-/// that is the buffer's `[pre…, grid…, tile…]` order, with each untiled axis dropping out after
-/// the coarsest level. Per-axis counts rather than a start/depth pair, so a buffer whose axes are
-/// tiled to different depths needs no new shape of description.
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct StorageTiling {
-    fragments: SmallVec<[usize; MAX_AXES]>,
-}
-
-impl StorageTiling {
-    /// Every axis split `levels + 1` ways. `levels = 0` is the untiled projection.
-    pub fn uniform(rank: usize, levels: usize) -> Self {
-        StorageTiling::per_axis(&vec![levels + 1; rank])
-    }
-
-    /// Axes from `start_axis` on split `levels + 1` ways, the ones before it untiled: a buffer
-    /// carrying batch (or otherwise passthrough) axes ahead of its tiled block.
-    pub fn suffix(rank: usize, start_axis: usize, levels: usize) -> Self {
-        assert!(
-            start_axis <= rank,
-            "StorageTiling::suffix: start_axis {start_axis} past rank {rank}"
-        );
-        StorageTiling::per_axis(
-            &(0..rank)
-                .map(|i| if i < start_axis { 1 } else { levels + 1 })
-                .collect::<Vec<_>>(),
-        )
-    }
-
-    /// An explicit fragment count per axis, in the operand's axis order.
-    pub fn per_axis(fragments: &[usize]) -> Self {
-        assert!(
-            fragments.iter().all(|&n| n >= 1),
-            "StorageTiling: every axis carries at least one physical fragment, got {fragments:?}"
-        );
-        StorageTiling {
-            fragments: SmallVec::from_slice(fragments),
-        }
-    }
-
-    pub fn rank(&self) -> usize {
-        self.fragments.len()
-    }
-
-    /// The number of physical fragments the axis at logical position `i` is split across.
-    pub fn fragments(&self, i: usize) -> usize {
-        self.fragments[i]
-    }
-
-    /// How many levels the physical order runs over: the deepest axis's fragment count.
-    pub fn depth(&self) -> usize {
-        self.fragments.iter().copied().max().unwrap_or(0)
-    }
-
-    /// The physical rank this induces, which is the buffer's own rank.
-    pub fn physical_rank(&self) -> usize {
-        self.fragments.iter().sum()
-    }
-
-    /// Whether any axis is split at all.
-    pub fn is_tiled(&self) -> bool {
-        self.fragments.iter().any(|&n| n > 1)
-    }
-
-    /// The physical axis labels this tiling induces over `axes`, in buffer order: the level-major
-    /// emission itself, one entry per physical axis, a tiled axis appearing once per fragment. The
-    /// single place the order is defined, shared by [`Projection::tiled`] and by callers labeling a
-    /// binding's dims ([`ConcreteLayout`]).
-    pub fn order(&self, axes: &[Axis]) -> Vec<Axis> {
-        assert_eq!(
-            self.rank(),
-            axes.len(),
-            "StorageTiling::order: the tiling describes {} axes but {} were given",
-            self.rank(),
-            axes.len()
-        );
-        let mut order = Vec::with_capacity(self.physical_rank());
-        for level in 0..self.depth() {
-            for (i, &axis) in axes.iter().enumerate() {
-                if level < self.fragments(i) {
-                    order.push(axis);
-                }
-            }
-        }
-        order
-    }
-}
+use crate::{Axis, ConcreteLayout, MAX_AXES, PhysicalAxisMap, Scale, StorageTiling};
 
 /// An operand's logical axes mapped onto its buffer's physical axes.
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
@@ -514,100 +304,6 @@ impl Projection {
         );
     }
 }
-
-/// The logical extent per axis, folded from `projection`'s physical shape: a single-carrier axis
-/// passes its physical extent through, a storage-tiled one multiplies its fragments' extents back
-/// together. Reduces to `physical_shape` itself for an untiled projection. A free function, not a
-/// method: `Projection` stays comptime-only (never a [`CubeType`]), so the runtime folding it
-/// drives lives in an ordinary `#[cube]` function, like every other comptime/runtime mix in this
-/// crate ([`top_window`](crate::GmemLayout)).
-#[cube]
-pub fn logical_extent(
-    #[comptime] projection: Projection,
-    physical_shape: &Coords<u32>,
-) -> Coords<u32> {
-    let mut bound = Coords::<u32>::new();
-    #[unroll]
-    for i in 0..comptime!(projection.logical_rank()) {
-        let picks = comptime!(projection.carriers(projection.axes[i]).to_vec());
-        bound.push(physical_shape.fproduct(picks));
-    }
-    bound
-}
-
-/// The line offset one `edge`-sized tile step along `axis` moves under `projection`: `axis`'s
-/// digits taken of `edge` itself rather than of a coordinate, dotted with `strides`. Exact because
-/// one edge-step's offset is linear in the tile index: every edge divides its enclosing block, so
-/// decomposing the step size the same way as a coordinate reconstructs the same advance. The
-/// radices come from `physical_shape`, so a static store folds the whole thing to a constant and a
-/// runtime-shaped one stays exact.
-#[cube]
-pub fn step_offset(
-    #[comptime] projection: Projection,
-    #[comptime] axis: Axis,
-    #[comptime] edge: usize,
-    physical_shape: &Coords<u32>,
-    strides: &Coords<u32>,
-) -> u32 {
-    let carriers = comptime!(projection.carriers(axis));
-    let picks = comptime!((0..carriers.len()).collect::<Vec<_>>());
-    let mut parts = Sequence::<u32>::new();
-    #[unroll]
-    for k in 0..comptime!(carriers.len()) {
-        let pa = comptime!(carriers[k]);
-        let (finer, modulo) = comptime!(projection.digit(pa, axis));
-        let scale = comptime!(projection.scale(pa, axis) as u32);
-        let quot = comptime!(edge as u32)
-            .runtime()
-            .fdiv(physical_shape.fproduct(comptime!(finer.to_vec())));
-        let digit = match comptime!(modulo) {
-            Some(m) => quot.frem(physical_shape.at(m)),
-            None => quot,
-        };
-        parts.push(digit.fmul(comptime!(scale).runtime()).fmul(strides.at(pa)));
-    }
-    parts.fsum(picks)
-}
-
-/// The inverse of [`GmemLayout::to_source_pos`](crate::GmemLayout::to_source_pos): the logical
-/// coordinate under `projection` that produced physical digits `digits` (one entry per physical
-/// axis, already decoded off the flat physical index). Requires
-/// [`Projection::is_invertible`]; a gathered (affine, scale != 1) projection never reaches this,
-/// since a stage is always materialized dense through [`Projection::direct_over`] first.
-#[cube]
-pub fn fold_physical(
-    #[comptime] projection: Projection,
-    digits: &Coords<u32>,
-    physical_shape: &Coords<u32>,
-) -> CoordsDyn {
-    comptime!(assert!(
-        projection.is_invertible(),
-        "Projection::fold_physical: not invertible (an affine term mixes several logical \
-         coordinates into one physical cell)"
-    ));
-    let mut out = CoordsDyn::new();
-    #[unroll]
-    for i in 0..comptime!(projection.logical_rank()) {
-        let axis = comptime!(projection.axes[i]);
-        let carriers = comptime!(projection.carriers(axis));
-        let picks = comptime!((0..carriers.len()).collect::<Vec<_>>());
-        let mut parts = Sequence::<u32>::new();
-        #[unroll]
-        for k in 0..comptime!(carriers.len()) {
-            let pa = comptime!(carriers[k]);
-            // Each digit weighs the block it sits above, the same finer extents that stripped it.
-            let (finer, _) = comptime!(projection.digit(pa, axis));
-            parts.push(
-                digits
-                    .at(pa)
-                    .fmul(physical_shape.fproduct(comptime!(finer.to_vec()))),
-            );
-        }
-        out.push(parts.fsum(picks));
-    }
-    out
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -792,37 +488,6 @@ mod tests {
         assert_eq!(p, Projection::of_layout(&realized));
     }
 
-    /// Folding `Σ digit * (extents it stripped)` back reconstructs the coordinate the digits were
-    /// decomposed from, which is what `fold_physical` does to `to_source_pos`. Both are `#[cube]`,
-    /// so the round trip is checked here on the digit positions they are built from.
-    #[test]
-    fn fold_physical_digits_invert_to_source_pos_digits() {
-        use crate::PhysicalAxis;
-
-        let shape = [3usize, 8];
-        let layout = ConcreteLayout::new(&[
-            PhysicalAxis::new(A, shape[0]),
-            PhysicalAxis::new(A, shape[1]),
-        ]);
-        let p = Projection::of_layout(&layout);
-        assert!(p.is_invertible());
-        let block = |pa| {
-            p.digit(pa, A)
-                .0
-                .iter()
-                .map(|&q| shape[q])
-                .product::<usize>()
-        };
-
-        for coord in [0usize, 1, 7, 8, 9, 19, 23] {
-            let digit = |pa| match p.digit(pa, A).1 {
-                Some(m) => (coord / block(pa)) % shape[m],
-                None => coord / block(pa),
-            };
-            assert_eq!(digit(0) * block(0) + digit(1) * block(1), coord);
-        }
-    }
-
     /// A spec built from a realized tiled layout is honest about its buffer: its physical rank
     /// *is* the tensor's rank, which is what `Tile::of` reads its shape and strides over, and its
     /// positional relabeling is the synthetic map the layout is addressed through. The declared
@@ -877,28 +542,6 @@ mod tests {
         assert_eq!(
             p.positional(),
             Projection::of_tiling(StorageTiling::uniform(2, 0))
-        );
-    }
-
-    /// The uniform and suffix tilings are the two the engine builds today; both are special cases
-    /// of the per-axis counts, and a zero-level tiling is `direct` itself.
-    #[test]
-    fn tiling_shorthands_agree_with_explicit_counts() {
-        assert_eq!(
-            StorageTiling::uniform(3, 1),
-            StorageTiling::per_axis(&[2, 2, 2])
-        );
-        assert_eq!(
-            StorageTiling::suffix(3, 1, 1),
-            StorageTiling::per_axis(&[1, 2, 2])
-        );
-        assert_eq!(
-            StorageTiling::suffix(2, 2, 3),
-            StorageTiling::per_axis(&[1, 1])
-        );
-        assert_eq!(
-            Projection::tiled(&[A, B], StorageTiling::uniform(2, 0)),
-            Projection::direct(&[A, B])
         );
     }
